Very early 3D, analytical solution interp err minimization, improved back mesh seeding (when using -back) and inveval HO.
- Early 3D: tetrahedra cavity operator. Mid-level functions: insertion, collapse, surface swap. High-level functions (insert long, collapse short, swap all) in 3D. Generalized smoothing (interior), localization to 3D. Fixed bugs related to 3D in various functions. Note: very little testing (e.g. only with CAD and one geometry), not expected to be robust, fast or yield good results.
- Interpolation error minimization of an analytical function. This will smooth the mesh to attempt to minimize a pnorm of Lagrange interpolation error. Works for HO and linear meshes. Reduction factors of 6x+ observed in limited testing (2D, anafun 5). This is meant for validating optimal metrics and metric smoothing approaches for HO meshing in the future, and overall motivation for curvature as an adaptation driver beyond anisotropy.
	- -anasol (same as -anamet) specifies analytical solution, implementations are in anasol2D.cxx / anasol3D.cxx.
	- -intp_pnorm and -intp_pdeg specify L^p norm and Lagrange interpolation degree respectively. 1-2 instantiated.
	- -smoo-type 1 applies solution-based smoothing instead of metric-based
	- Two optimizers, one based on Newton, one on an adaptation of DIRECT (D. R. Jones, "Lipschitzian optimization without the Lipschitz constant") to groups of triangles/tetrahedra. The latter does global optimization, it is very expensive but finds optima better than we can hope for using local optimization, for the purpose of testing this error minimization approach. Currently, we call the former, then Newton-based to ensure a tight optimum (this ends up never doing anything, but at least we know DIRECT went as far as we could hope for).
	- Only tested in 2D. DIRECT has been tested in 3D but needs considerable improvements (the splits are no good, or there is a bug).
- Refactored and improved interpFrontBack to seed points from nothing, starting from corners, linear complexity (no octree). Limited testing with Q2 back meshes (-back option). Improved localization in Q2 meshes, namely luksan_pnet had some hidden stopping criteria preventing from hitting the requested tolerances.
- Improved interpMetBack, when a point is sufficiently initialized (e.g. is a vertex in the mesh), a simple msh.interpMetBack(ipoin) can be used. A point can now also be seeded by another point in the mesh, rather than an element. Also added MeshBase::get_algnd() for such points.
- Cavity no longer needs a normal from the caller, it computes its own per connex component (by ref) of the surface cavity.
- Removed all default ithread arguments (fixing a few bugs), this needs refactoring in the future (MeshBase should have a get_itag() and handle availability of ithread/itags on its own).

+++ tools/gen_lag2bezfuncs.cxx
@@ -261,7 +261,7 @@ void gen_lag2bez(std::ostringstream &str){
     
         if(szfld > 1){
           str << "    rfld1[lfld["<<i_s<<"]][i] = "<<p<<"*rfld0[lfld["<<j_s<<"]][i]/"<<q;
-          for(int j = 1; j < pcoefs[i].size(); j++){
+          for(int j = 1; j < (int) pcoefs[i].size(); j++){
             #ifndef USE_DBL_COEFS
               p = pcoefs[i][j];
               q = qcoefs[i][j];
@@ -274,7 +274,7 @@ void gen_lag2bez(std::ostringstream &str){
           }
         }else{
           str << "  rfld1[lfld["<<i_s<<"]][0] = "<<p<<"*rfld0[lfld["<<j_s<<"]][0]/"<<q;
-          for(int j = 1; j < pcoefs[i].size(); j++){
+          for(int j = 1; j < (int) pcoefs[i].size(); j++){
             #ifndef USE_DBL_COEFS
               p = pcoefs[i][j];
               q = qcoefs[i][j];
